修改商品默认物流价格的bug

+++ app/src/main/java/com/thn/lexi/order/SubmitOrderBean.java
@@ -173,7 +173,7 @@ public class SubmitOrderBean {
                 public String rid;
                 public String s_color;
                 public String s_model;
-                public int s_weight;
+                public double s_weight;
                 public int sale_price;
                 public int stock_count;
             }
